// web-ANGULAR/dfps-web-impact/src/app/modules/case/model/homeAssessmentDetail.ts
export interface DisplayHomeAssessmentDetail {
    homeAssessmentId: number;
    assessmentStatus: string;
    approvalStatusHide: boolean;
    criminalHistoryPending: boolean;
    criminalHistoryPersonList: any[];
    documentExists: boolean;
    documentProtected: boolean;
    enabled: boolean;
    event: Event;
    eventType: string;
    eventTypeCode: string;
    homeAssessmentHistory: HomeAssessmentHistory[];
    lastUpdatedDate: string;
    outputCompletedChecked: boolean;
    outputCompletedEnabled: boolean;
    overallDate: string;
    overallStatus: string;
    pageMode: string;
    radioBtnEnabled: boolean;
    radioChecked: boolean;
    saveHidden: boolean;
    signatureDate: string;
    stageId: number;
    stageName: string;
    statusCodes: DropDown[];
    statusHidden: boolean;
    submitHidden: boolean;
    submittedForApproval: boolean;
    signatureFieldsEnabled: boolean;
    attentionMessages: string[];
    approvalMode: boolean;
    formTagDto: FormTagDto;
    showInvalidateApprovalPopup: boolean;
    showInitialApproval : boolean;              // added for artf246288
}
export interface Event {
    caseId: number;
    createdDate: string;
    description: string;
    eventId: number;
    lastUpdatedDate: string;
    modifiedDate: string;
    occurredDate: string;
    personId: number;
    purgedDate: string;
    stageId: number;
    status: string;
    taskCode: string;
    type: string;
}
export interface HomeAssessmentHistory {
    eventId: number;
    fullName: string;
    id: string;
    lastUpdatedDate: string;
    statusChangedDate: string;
    statusCode: string;
}

export interface FormTagDto {
    action: string;
    approvalFlag: string;
    attachmentFormName: string;
    bFormContact: string;
    bPreview: string;
    bResend: string;
    ccEmailAddress: string;
    cdContactMethod: string;
    cdContactOthers: string;
    cdContactPurpose: string;
    cdRelation: string;
    cdStage: string;
    checkForNewMode: string;
    checkStage: string;
    cpsContractAddendum: string;
    deleteDocument: string;
    displayName: string;
    displayOnly: string;
    docExists: string;
    docType: string;
    dtContactOccurred: string;
    dtTimestamp: string;
    emailContent: string;
    emailFrom: string;
    emailNotify: string;
    emailTitle: string;
    emailToList: string;
    eventType: string;
    formName: string;
    hdnCalledFromJavascript: string;
    hdnFlagcontractaddendum: string;
    hdnFlagcpsaddendum: string;
    hiddenField: string;
    hideInWidget: string;
    id: string;
    idCase: string;
    idCreatedPerson: string;
    idEvent: string;
    idEventReq: string;
    idEventTable: string;
    idFamilyPlanEvalEvent: string;
    idFamilyPlanEvent: string;
    idFormReferral: string;
    idFsnaEvent: string;
    idICPCTransmittal: string;
    idLastUpdatePerson: string;
    idLegalNoticeDtl: string;
    idLegalNoticeRecpnt: string;
    idMedCons: string;
    idPPTEvent: string;
    idPPTPart: string;
    idPerson: string;
    idPersonTable: string;
    idPrmryPrntCrgvr: string;
    idReferral: string;
    idResource: string;
    idResourceTable: string;
    idRiskAsmntLkp: string;
    idSSCCPlcmtHeader: string;
    idScndryPrntCrgvr: string;
    idStage: string;
    idSvcPlnEvalEvent: string;
    idTransmittal: string;
    indAddOrModify: string;
    level1Tab: string;
    level2Tab: string;
    level3Tab: string;
    modeOfPage: string;
    nCase: string;
    nEvent: string;
    name: string;
    nbrVersion: string;
    nmStage: string;
    onClick: string;
    pArClosureReason: string;
    pCase: string;
    pCdNotifStat: string;
    pCdStage: string;
    pChild: string;
    pCrimHist: string;
    pDtCPSInvstDtlBegun: string;
    pDtLicngInvstDtlBegun: string;
    pDtSampleFrom: string;
    pDtSampleTo: string;
    pEvent: string;
    pFormName: string;
    pIcpcRequest: string;
    pMedCons: string;
    pParameter: string;
    pPerson: string;
    pPersonCitizenship: string;
    pPersonId: string;
    pPptEvent: string;
    pRecCheck: string;
    pRecordsCheckNotif: string;
    pRegion: string;
    pRejection: string;
    pStage: string;
    pStageReasonClosed: string;
    pSzCdEventType: string;
    pSzCdStageType: string;
    pWorker: string;
    postInSameWindow: string;
    promptNoDocument: string;
    promptSavePage: string;
    protectDocument: string;
    renderFormat: string;
    sCase: string;
    sEvent: string;
    sIdRecCheck: string;
    sIdTletsCheck: string;
    sRecordsCheckNotif: string;
    sStage: string;
    sTimestamp: string;
    selCdSearchType: string;
    selSzCdStageReasonClosed: string;
    selectedEpisodes: string;
    switchLanguage: string;
    taskCD: string;
    timestamp: string;
    tmScrTmCntct: string;
    txtNmPersonFull: string;
    userId: string;
    userIdPerson: string;
    userName: string;
    windowName: string;
    xpandContactSearch: string;
    xpandGlobalData: string;
    pResource: string;
    newTimestamp: string;
}